Corrigido bug com Plot
Add implementação para não permitir mover componentes quando conectados.

+++ Electrophorus.Rendering/Windows/AboutCapacitor.cs
@@ -26,7 +26,7 @@ namespace Electrophorus.Rendering.Windows
         {
             if (txtCapacitancia.Text != string.Empty)
             {
-                _capacitor.capacitance = int.Parse(txtCapacitancia.Text);
+                _capacitor.capacitance = double.Parse(txtCapacitancia.Text);
             }
             if (View != null) View.Refresh();
             Close();
